Make Group.js a unique component using short-circuit evaluation
Remove dynamic child components
